feat: add sequelize backed by sqlite
Flesh out the backend further by introducing a Sequelize model
and necessary encapsulations to persist/retrieve checker instances.

Allow the user to specify the primary key, enforcing uniqueness. This
is in-line with how short urls are unique in GoGovSg and usable
as PKs; in a similar vein, we want urls to checkers to be memorable
to humans.

- Add sequelize, sqlite
- Rename the Config interface in types/checker to Checker
  - Add associated Sequelize model
- Create a service wrapper around Checker model, injecting all this
  into the controller
- Rewrite the controller to accommodate Sequelize (via the service)

TODO: tests, user model