feat(express): strap into serverless-http
Because we have to wait for `sequelize.sync()` before making the
Express app available, we cannot obtain a serverless handler directly
if we are going to use it in top-level module code, and have to access
it through a Promise. However, AWS requires us to expose the serverless
handler alone, and cannot unpack it from a Promise resolving to one.

To get around this, implement an AWS Lambda handler that will await
on the Promise that is creating the serverless handler from the
app returned by the asynchronous `bootstrap()`. In effect, we are
unwrapping the Promise on AWS' behalf.