feat(production): prepare for deployment (#66)
* refactor: move bootstrapping into separate module

Move most of the bootstrapping block in index.ts into its own module,
with a view to further dividing the block into specific modules for
sequelize, nodemailer and logging

* refactor(bootstrap): reorg into sequelize, mailer, logger

- Shard out bootstrap logic, creating
  sequelize, mailer and logger modules
- Allow nodeEnv to influence what the modules expose to bootstrap,
  and create sensible objects to use if nodeEnv is not development
- Take the trouble to tweak how Nodemailer's `sendMail()` is injected
  into AuthService, on learning that it can return a Promise
- Add morgan logging to express bootstrapping

* feat: inject logger into AuthController, CheckerService

- Accept logger as an optional dependency, logging useful messsages
  in lieu of console.{error, warn}
- Take the trouble to log unauthorized OTP attempts

* feat(express): strap into serverless-http

Because we have to wait for `sequelize.sync()` before making the
Express app available, we cannot obtain a serverless handler directly
if we are going to use it in top-level module code, and have to access
it through a Promise. However, AWS requires us to expose the serverless
handler alone, and cannot unpack it from a Promise resolving to one.

To get around this, implement an AWS Lambda handler that will await
on the Promise that is creating the serverless handler from the
app returned by the asynchronous `bootstrap()`. In effect, we are
unwrapping the Promise on AWS' behalf.