fix(auth): change user localStorage to User | null (#222)
`useLocalStorage()` uses `JSON.parse()` to retrieve previously
stringified values from localStorage. This is problematic as
`JSON.stringify()` serializes `undefined` to the string "undefined",
which it is then unable to parse.

To mitigate this, change the type of `[user, setUser]` to be
`User | null`, so that we are forced to avoid `undefined`. This also
better lines up with what is returned by `/auth/whoami` in the backend.